Disable TSan reporting on parts of the runtime
TSan warns about data races on these functions. We reported those
warnings in #11040, and silence them to avoid facing users of
TSan with data race reports that are not related to their code.
This is done in two ways: either un-instrumenting those functions, or
adding their name in __tsan_default_suppressions in tsan.c.

Co-authored-by: Fabrice Buoro <fabrice@tarides.com>