Build system: do not use SHAREDLIB_CFLAGS when computing dependencies
This commit containstwo changes to the root Makefile:

1. Removal of the following line:

$(DEPDIR)/runtime/%.bpic.$(D): OC_CFLAGS += $(SHAREDLIB_CFLAGS)

This line is useless because it makes no sense to alter any variable
in the CFLAGS category for the computation of dependencies, since only
the C preprocessor is involved in this stage and it does not take
C flags into account anyway, only C preprocessor flags.

2. When computing the dependencies for $(DEPDIR)/runtime/%.npic.$(D),
one should not refer to $(SHAREDLIB_CFLAGS), for a similar reason.

It has been verified that SHAREDLIB_CFLAGS is either empty, or contains
just -fPIC which is indeed not necessary for computing dependencies
(it is indeed a C flag rather than a C preprocessor flag).